refactor(state): isolate database cache lifecycle
